Reorganize models into categories
