+++ src/backend/src/main/java/com/prepple/api/controller/QuestionController.java
@@ -8,8 +8,8 @@ import com.prepple.api.configuration.Authorization;
 import com.prepple.api.configuration.Route;
 import com.prepple.api.configuration.ServiceConfig;
 import com.prepple.api.dto.QuestionDto;
-import com.prepple.api.model.Question;
-import com.prepple.api.model.QuestionBatchRequest;
+import com.prepple.api.model.questions.Question;
+import com.prepple.api.model.questions.QuestionBatchRequest;
 import com.prepple.api.service.AuthService;
 import com.prepple.api.service.QuestionService;
 import com.prepple.api.service.SessionService;
