+++ src/backend/src/main/java/com/prepple/api/dto/LabelAssignmentDto.java
@@ -1,6 +1,6 @@
 package com.prepple.api.dto;
 
-import com.prepple.api.model.LabelAssignment;
+import com.prepple.api.model.questions.LabelAssignment;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Data;
