+++ src/backend/src/main/java/com/prepple/api/model/questions/LabelType.java
@@ -1,4 +1,4 @@
-package com.prepple.api.model;
+package com.prepple.api.model.questions;
 
 /**
  * Types of labels that can exist in the system
