+++ src/backend/src/main/java/com/prepple/api/model/shared/JwtPayload.java
@@ -1,4 +1,4 @@
-package com.prepple.api.model;
+package com.prepple.api.model.shared;
 
 import lombok.Data;
