+++ src/backend/src/main/java/com/prepple/api/util/JwtDecoder.java
@@ -3,9 +3,9 @@ package com.prepple.api.util;
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.databind.ObjectMapper;
 import com.prepple.api.configuration.ServiceConfig;
-import com.prepple.api.model.Jwt;
-import com.prepple.api.model.JwtHeader;
-import com.prepple.api.model.JwtPayload;
+import com.prepple.api.model.shared.Jwt;
+import com.prepple.api.model.shared.JwtHeader;
+import com.prepple.api.model.shared.JwtPayload;
 import io.jsonwebtoken.SignatureAlgorithm;
 import io.jsonwebtoken.impl.crypto.DefaultJwtSignatureValidator;
 import org.springframework.beans.factory.annotation.Autowired;
