+++ src/backend/src/test/java/com/prepple/api/dto/MediaFileDtoTest.java
@@ -1,6 +1,5 @@
 package com.prepple.api.dto;
 
-import com.prepple.api.model.LabelType;
 import org.junit.jupiter.api.Test;
 
 import java.lang.reflect.Method;
